requirements.txt
entsoe-py>=0.7.7,<0.8.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
streamlit>=1.43,<2.0
plotly>=6.0,<7.0
openpyxl>=3.1,<4.0

# Dev dependencies
pytest>=8.0,<9.0
